+++ mrb_doc/web.rb
@@ -1,6 +1,7 @@
 # Sets the main loop for web exports
 # @param method [String]
 # @return [nil]
+# @raise [PlatformSpecificMethodCalledOnWrongPlatformError] If run on the wrong platform
 def set_main_loop(method)
   # mrb_set_main_loop
   # src/web.cpp