+++ test/fixtures/web.rb
@@ -1,0 +1,5 @@
+$set_main_loop_run = false
+
+def set_main_loop_stub
+  $set_main_loop_run = true
+end